Kafka-2992: Guard trace statements in the inner loop of the replica fetcher
We're seeing some GC pause issues in production, and during our investigation found that the thunks created during invocation of three trace statements guarded in this PR were responsible for ~98% of all allocations by object count and ~90% by size. While I'm not sure that this was actually the cause of our issue, it seems prudent to avoid useless allocations in a tight loop.

I realize that the trace() call does its own guarding internally, however it's insufficient to prevent allocation of the thunk.

This is my original work, and I license it to the Kafka project under the project's Apache license.

Author: Cory Kolbeck <cory.kolbeck@urbanairship.com>

Reviewers: Gwen Shapira

Closes #682 from ckolbeck/guard-trace-statements