KAFKA-2146: Fix adding partition did not find the correct startIndex.
TopicCommand provide a tool to add partitions for existing topics. It try to find the startIndex from existing partitions. There's a minor flaw in this process, it try to use the first partition fetched from zookeeper as the start partition, and use the first replica id in this partition as the startIndex.
One thing, the first partition fetched from zookeeper is not necessary to be the start partition. As partition id begin from zero, we should use partition with id zero as the start partition.
The other, broker id does not necessary begin from 0, so the startIndex is not necessary to be the first replica id in the start partition.

Author: chenshangan <chenshangan@meituan.com>

Reviewers: Guozhang Wang

Closes #329 from shangan/trunk-KAFKA-2146