KAFKA-1860: Fix issue that file system errors are not detected unless Kafka tries to write.
When the disk (raid with caches dir) dies on a Kafka broker, typically the filesystem gets mounted into read-only mode, and hence when Kafka tries to read the disk, they'll get a FileNotFoundException with the read-only errno set (EROFS). However, as long as there is no produce request received, hence no writes attempted on the disks, Kafka will not exit on such FATAL error and keep on throwing exception : java.io.FileNotFoundException

In this case, the JVM should stop if the underlying file system goes in to Read only mode.

Author: MayureshGharat <gharatmayuresh15@gmail.com>

Reviewers: Lin Dong, Gwen Shapira, Ismael Juma, Guozhang Wang

Closes #698 from MayureshGharat/KAFKA-1860