KAFKA-3111; Fix ConsumerPerformance reporting to use time-based instead of message-based intervals
Interval lengths for ConsumerPerformance could sometime be calculated as zero. In such cases, when the bytes read or messages read are also zero a NaN output is returned for mbRead per second or for nMsg per second, whereas zero would be a more appropriate output.

In cases where interval length is zero but there have been data and messages to read, an output of Infinity is returned, as expected.

Author: Vahid Hashemian <vahidhashemian@us.ibm.com>

Reviewers: Dong Lin <lindong28@gmail.com>, Jun Rao <junrao@gmail.com>

Closes #788 from vahidhashemian/KAFKA-3111