KAFKA-3945; Change the type of 'acks' config in console producer to String
The `acks` config that is provided to console producer with `request-required-acks` comes with `all`, `-1`, `0`, `1` as valid options (`all` and `-1` being interchangeable). Currently, the console producer expects an integer for this input and that makes `all` to become an invalid input. This PR fixes this issue by changing the input type to String.

Author: Vahid Hashemian <vahidhashemian@us.ibm.com>

Reviewers: Manikumar reddy O <manikumar.reddy@gmail.com>, Grant Henke <granthenke@gmail.com, Ismael Juma <ismael@juma.me.uk>

Closes #1618 from vahidhashemian/KAFKA-3945