MINOR: Update the README.md to include a note about GRADLE_USER_HOME
Trying to build the source and publish it to internal Maven repo, I ran into an issue that I explain in the mailing list discussion here https://www.mail-archive.com/devkafka.apache.org/msg56359.html.

The commit here updates the README.md to make a note that the GRADLE_USER_HOME environment variable plays a role in deciding which file to add the maven configs to.

Author: Jaikiran Pai <jaikiran.pai@gmail.com>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #1837 from jaikiran/readme-update-grade-user-home