KAFKA-4409; Fix deadlock between topic event handling and shutdown in ZookeeperConsumerConnector
The consumer can deadlock on shutdown if a topic event fires during shutdown. The shutdown acquires the rebalance lock and then the topic-event-watcher lock. The topic event watcher acquires these in the reverse order. Shutdown should not need to acquire the topic-event-watcher’s lock - all it does is unsubscribes from topic events.

Author: Joel Koshy <jjkoshy@gmail.com>

Reviewers: Jiangjie Qin <becket.qin@gmail.com>

Closes #2129 from jjkoshy/KAFKA-4409